requirements.txt
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
drf-yasg==1.20.0
